🍌🪓
Creating a top level c and a top level js
directory in order to track two different
lines of development in a single concurrent
place.

We are currently trying to get a working js
version that will run on top of Node. We sort
of had that but got distracted chasing
tangents. It needs to be cleaned up and
formalized, versioned and shipped so that
people can build on top of it and know that
the carpet isn't just going to be pulled out
from underneath of them.

Also working on a c version that will
hopefully be faster and probably most likely
be running on the web assembly. This is a
deep dive into c for me. I selected c because
it is still a language that fascinates me,
even though we really could have it written
in any language, and I suppose these are all
really just translations of a single idea or
principle, whatever that may be.

In order to give some hints, and what is
known about bubblescript, as i call it, is
really just a Lisp or Clojure that works like
Javascript. Kind of a loose wrapper or
replacement for it. It is not all that unique
a project in that regard. You could say it is
a programming language dialect, but my hope
is that there is some prime basis for it.
Some reasoning behind drilling through all
the superfcialities to find a programming
core the can ultimatly help to improve the
art in some fashion.

What is, and what isn't. All to be discovered
upon the path towards enlightenment.
Farewell!

Kandii Kangaroo 🍭🦘