In order to get a high
level view of what we are
working with i have
tentively organized the
files here into three sub
directories. At the root
we have the bubls.js files
which is the main file
that loads the project.
The f directory has files
which contains functions.
The o directory has the
class definitions for
obects and the z directory
has everything else.

Like i said, this is just
a temporary fix to get
barings, but i like how
clean it makes the break
down so will prolly keep
and build on this to keep
navogation and
understanding of
everything crystal clear.

🫏