🦝 There was bug in tokenize
because when we were checking for true in the regex we didnt have the
caret to make sure it was the first thing in the string. This has the
fix and is mostly just a memoize commit. Needs lots of clean, but fixed
the anyone bug that truned all symnols into true when the was a true
symbil present.