. map and empty? in bubblescript
This commit captures a first attempt a map doing it as a binding in
rootBinding and delegating to list. This works as long as ising a jsfn,
but is the wrong approach, as doing that is desirably avoided. Instead,
implementing map as a function works better, and keeps everything in
bubblescript, no need to map function. This wasn't possible with some of
the core functions, the glue functions, but having thosr inplace we
shpuldnow be able to start implementing new functionality in
bubblescript. Will need to devise a means to organize and run test, and
to break up and maintain and compose namespaces/modules, and already
thinking aboit best ways to do that.

Some of these commits are just experimental, amd will be cleaned up and
fleshed oit in future commits.

Cheers, BaMbi ♡♡