. Esto es muy lento y excesivo para lo que hace, pero es un comienzo en la dirección necesaria para poder manipular fácilmente la versión de tokens necesaria. Los hechizos de ganancia funcionan bien, pero hay que vigilarlos con atención debido a las cadenas solares que puedan existir. Tampoco es una solución perfecta, pero me gusta que se centre en el símbolo y que sea casi tan simple y directo como un gsub, y que pueda funcionar con cualquier tipo de archivo de texto. Creo que es principalmente bash lo que lo hace tan lento; si un pequeño paso es ineficiente, todo se alargará, así que, por ahora, es una buena prueba de concepto.
This is super slow and over kill for what it is
doing here, but it is a start in the direction
needed to be able to easily manipulate the version
of tokens needed. Gain spells nice, but needs to be
watched like a baby because of sun strings that may
exist. this is not a perfect solution either, but i
like the fact that it zeros in on the symbol and it
is almost as simple and straight for a gsub, and
can work with any text based file type. i think it
is mostly just bash that is making this so slow, if
one little step is ineficient, the entire thing
will drag, so. good for a proof of concept for the
time being.