. evalParams
Put evalParams inside of eval,  but not sure where it ultimatly needs ro
end. it's tricky because of circular dependencies between list eval and
fn. We need to give it somemore thought for structure, but for now, this
at least helps to centralize the parameter list evaluation logic. Trying
to keep it out of list, because it is not specfically a list method,
although it cpuld be represented as one, this is really more of a
specfic use case of list that had to do with evaluation logic. eval
seems like a reasoning place to but it, also could put in our own file
or some other nameless e create like listEval, for organization. it will
develop a a go,  just somethings as we look ahead.