☘️ Getting a pseudotest for definir trabojing. 🤸🏿‍♀️
This is a first pass, and since it's passing gonna commit. We are avoid
the need to load evalParams on init by passing it along with the call to
Fn.call. Much easier to visualize, understand and work with. Functional
win.

There still might be a cleaner way to have this all constructed (almost
certianly) but this emergent and the flexability providid is
an advantage as we step forward incrementally.

    bubblescript 🍀 npx mocha test/definir.js

      definir
      { 'vínculo': { define: [Function: define] } }
      Hola Mama!
      { 'vínculo': { define: [Function: define] } }
          ✔ sets a constant

            1 passing (45ms)