🪶 Ok, all three test for definir...
... are passing now, although the test coverage is kind of weak and the
last test i just added is a little wonky, but does prove pretty weel
that the function was added, although i can see how it could be a false
positive right now, but not likely. We had to inject more eval functions
to give the Fn methods everything the required. This can be simplified
by passing in the eval exported object, figuring that out in follow ups.

bubblescript 🍀 npx mocha test/definir.js

  definir
    ✔ establece un valor
    ✔ no define un valor si ya estaba definido (el valor es constante)
    ✔ creates and sets a function when psassed  a list as the first parameter, and uses the remainder of the list as the body.

  3 passing (53ms)