⚡️ slt-js.js
I had a discussion with ChatGPT regarding tge creation of a standard
mechinisim that would allow the usage and transpiling of localized
natural language keywords with a particular programming language. As a
quick shot and implementing a protype for demonstrational purposes,
ChatGPT prepared the slt-js.js file which is included in this commit.

I've used it, to transpile the file at prueba/símbolos/fn.es.js, which
is a file i've been working in most recently, although kind of a pale
example, since it doesn't use many keywords, to transpile to
prueba/símbolos/fn.js, and have good results, which are included.

Thinking a bit about directory structure in supporting more lanuages for
source.

Also have a more advance npm package which will be publishing shortly,
which, with better tokenization and parsing.

  bubblescript 🍀 node slt-js.js prueba/símbolos/fn.es.js -o
  prueba/símbolos/fn.js
  Wrote transpiled output to prueba/símbolos/fn.js