. make builds all files directly under fuente by ...
...running them through slt-js.js and saves output to the same file name
under build/es. Only processing top level files right now, neex to
modify command to work recursive, was also having an issue with the code
as it was using the COMMAND variable so had to inline it to get it
working. Still getting some weirdness with apoya, where it trys to pass
the directory to slt-js and crashes, but works fine if you run make
againg. Time stamps are working as expected as does clean. A good first
step.