. Quick pass at simplified require.
This is a different design for our require method that i had envisioned
to work with a frozen root binding object, however, we seem to miss the
mark on that one just a little. We need to do a little furthure
investigation. i guess it is because we would be shadowing the value on
the root object which is frozen, which is why the value, __dir name
can't be set on an object with the root binding as the prototype.
Anyway, will finish that up in a followup. This bring some other
simplification, we don't need to curry and all of that jazz, everything
just relies on __dir name, and we update that in require. All tests are
passing, and i believe that this gives us everything we need, or get's
us in the right directuon at least for the time being.